+++ public/admin/_app/immutable/chunks/BXZ4Q3d1.js
@@ -1,0 +1,1 @@
+import{l as e}from"./Du8n6nmR.js";function t(t,n){throw new e(t,n.toString())}export{t};