+++ public/admin/_app/immutable/chunks/onuViHCd2.js
@@ -1,1 +1,0 @@
-import{o as e,s as t}from"./Dqbbk3fc.js";import"./xihTtKlq.js";import{t as n}from"./BiY6hucH.js";function r(r,i){let a=e(i,[`$$slots`,`$$events`,`$$legacy`]),o=[[`path`,{d:`M12 19h8`}],[`path`,{d:`m4 17 6-6-6-6`}]];n(r,t({name:`terminal`},()=>a,{get iconNode(){return o}}))}export{r as t};